+++ CHANGES.md
@@ -17,6 +17,7 @@
 - Add `Yocaml.Data.Validation.String`, a set of validator for `String` (by [Okhuomon Ajayi](https://github.com/six-shot))
 - Add missing test coverage for `Nel.equal` and `Nel.append` functions (by [Bill Njoroge](https://github.com/Bnjoroge1))
 - Add `to_data` and `from_data` for Archetypes  (by [gr-im](https://github.com/xvw))
+- Add `Yocaml.Data.Validation.Int` and `Float`, a set of validator for `Int` and `Float`  (by [gr-im](https://github.com/gr-im))
 
 
 ### v2.6.0 2025-09-23 Nantes (France)